Add files via upload
Here we use a GMNO hypergraph embedding to produce an embedding from [0,1]\times[0,1] to 9 dimensional Euclidean space. This embedding can be used as a positional embedding for image processing and computer vision. It is easier to process a position in an image when that image is represented as a vector of length 9 than as a vector of length 2.